removed plotting routine from nautilus to generalize to abstract_searcxh
