+++ autofit/non_linear/search/nest/nautilus/search.py
@@ -557,32 +557,3 @@ class Nautilus(abstract_nest.AbstractNest):
             **config_dict,
             "n_like_max": 1,
         }
-
-    def plot_results(self, samples):
-        from autofit.non_linear.search.nest.nautilus.plotter import NautilusPlotter
-
-        if not samples.pdf_converged:
-            return
-
-        def should_plot(name):
-            return conf.instance["visualize"]["plots_search"]["nautilus"][name]
-
-        plotter = NautilusPlotter(
-            samples=samples,
-            output=Output(path=self.paths.image_path / "search", format="png"),
-        )
-
-        if should_plot("cornerplot"):
-            plotter.cornerplot(
-                panelsize=3.5,
-                yticksize=16,
-                xticksize=16,
-                bins=20,
-                plot_datapoints=False,
-                plot_density=False,
-                fill_contours=True,
-                levels=(0.68, 0.95),
-                labelpad=0.02,
-                range=np.ones(samples.model.total_free_parameters) * 0.999,
-                label_kwargs={"fontsize": 24},
-            )
